backend/requirements.txt
# Django and REST Framework
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
drf-spectacular==0.28.0

# Firebase
firebase-admin==6.5.0

# Machine Learning
scikit-learn==1.6.0
numpy==2.2.1
pandas==2.2.3
joblib==1.4.2

# PyTorch
torch==2.5.1
torchvision==0.20.1

# LangChain and Gemini AI
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-community==0.3.13
langchain-core==0.3.28

# Caching and Task Queue
redis==5.2.1
django-redis==5.4.0
celery==5.4.0

# Environment management
python-decouple==3.8

# Development and testing
pytest==8.3.4
pytest-django==4.9.0
hypothesis==6.122.2

# Security and utilities
cryptography==44.0.0
requests==2.32.3